armnn-onnx: use stable version 1.6.0 and repair build
Version previously indicated in this recipe was not the stable one (no
tag applied) and rather a snapshot of the tip from development branch.

Resolve this by using a stable tag for version 1.6.0 and wait until
1.6.1 gets it's tag applied.

In addition, indicate that protobuf3 should be used for build, exclude
-ml namespace from the build and repair installation by verifying that
only files are getting installed (no folders are taken).

+++ recipes-libraries/armnn/armnn-onnx_git.bb
@@ -2,12 +2,12 @@ SUMMARY = "ONNX protobuf files - used in ARMNN for Caffe network models"
 LICENSE = "MIT"
 LIC_FILES_CHKSUM = "file://LICENSE;md5=efff5c5110f124a1e2163814067b16e7"
 
-BRANCH = "rel-1.6.1"
-PV = "1.6.1"
+BRANCH = "rel-1.6.0"
+PV = "1.6.0"
 
 SRC_URI = "git://github.com/onnx/onnx.git;branch=${BRANCH}"
 
-SRCREV = "1facb4c1bb9cc2107d4dbaf9fd647fefdbbeb0ab"
+SRCREV = "553df22c67bee5f0fe6599cff60f1afc6748c635"
 
 DEPENDS = "protobuf-native"
 
@@ -15,13 +15,18 @@ inherit cmake
 
 S = "${WORKDIR}/git"
 
-EXTRA_OECMAKE += "-DPYTHON_EXECUTABLE=${HOSTTOOLS_DIR}/python3"
+EXTRA_OECMAKE += "-DPYTHON_EXECUTABLE=${HOSTTOOLS_DIR}/python3 -DONNX_VERIFY_PROTO3=ON -DONNX_ML=OFF"
 
+# Override the install task, since we are not interested in any of the libraries
+# delivered from the build and rather require to provide a Protobuf sources
 do_install() {
     install -d ${D}${datadir}/${BPN}/onnx/
     for file in ${B}/onnx/*
     do
-        install -m 0644 $file ${D}${datadir}/${BPN}/onnx/
+        if [ -f $file ]; then
+            install -m 0644 $file ${D}${datadir}/${BPN}/onnx/
+        fi
     done
 
 }
+
